+++ example/package.json
@@ -4,7 +4,7 @@
   "main": "index.js",
   "license": "MIT",
   "scripts": {
-    "test": "ts-node ./dockest.ts"
+    "test": "ts-node -r tsconfig-paths/register ./dockest.ts"
   },
   "dependencies": {
     "dotenv": "^6.2.0",
