+++ package.json
@@ -28,6 +28,7 @@
     "sequelize": "^4.41.2",
     "sequelize-cli": "^5.3.0",
     "ts-jest": "^23.10.5",
+    "tsconfig-paths": "^3.8.0",
     "tslint": "^5.11.0",
     "tslint-config-prettier": "^1.17.0",
     "tslint-eslint-rules": "^5.4.0",
