+++ src/runners/execs/teardown.ts
@@ -1,5 +1,5 @@
 import { ExecOpts } from '../index'
-import { teardownSingle } from './index'
+import { teardownSingle } from '../utils'
 
 export default async (execConfig: ExecOpts) => {
   const { containerId, runnerKey } = execConfig
