+++ tslint.json
@@ -15,7 +15,8 @@
     "no-duplicate-variable": true,
     "max-line-length": false,
     "interface-over-type-literal": false,
-    "no-implicit-dependencies": [true, "dev"]
+    "no-submodule-imports": [true, ["src"]],
+    "no-implicit-dependencies": [true, ["src"]]
   },
   "rulesDirectory": [],
   "linterOptions": {
